pyproject.toml
[tool.poetry]
name = "aries_endorser_service"
version = "0.0.1"
description = "Hyperledger Aries Endorser Service is a controller for an Aries Cloudagent Python agent acting as endorser."
authors = ["Hyperledger Aries <aries@lists.hyperledger.org>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "endorser"}]
classifiers = [
            "Programming Language :: Python :: 3",
            "License :: OSI Approved :: Apache Software License",
	        "Operating System :: OS Independent",
]
repository = "https://github.com/hyperledger/aries-endorser-service"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
alembic = "^1.13.1"
asyncpg = "^0.30.0"
fastapi = {version = "^0.115.8", extras = ["standard"]}
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
python-jose = {version = "^3.3.0", extras = ["cyptography"]}
requests = "^2.32.3"
SQLAlchemy = {version = "^2.0.30", extras = ["asyncio"]}
sqlmodel = "0.0.19" # see https://github.com/fastapi/sqlmodel/discussions/1048 before upgrading
starlette-context = "^0.3.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "~3.8.0"
# Sync with version in .pre-commit-config.yaml
ruff = "^0.9.7"
# Sync with version in .github/workflows/blackformat.yml
# Sync with version in .pre-commit-config.yaml
black = "^24.4.2"

# testing
behave = "^1.2.6"
pytest = "^8.2.0"
pytest-asyncio = "^0.25.3"
pytest-cov = "^5.0.0"
pytest-ruff = "^0.4.1"


[tool.poetry.extras]


[tool.poetry.scripts]
aca-py = "endorser.main:main"

[tool.ruff]
lint.select = ["B006", "C", "D", "E", "F"]

lint.ignore = [
    # Google Python Doc Style
    "D203", "D204", "D213", "D215", "D400", "D401", "D404", "D406", "D407",
    "D408", "D409", "D413",
    "D202", # Allow blank line after docstring
    "D104", # Don't require docstring in public package
    # Things that we should fix, but are too much work right now
    "D417", "C901",
]

include = ["endorser/**/*.py"]

line-length = 90

exclude = ["migrations"]
[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["B006", "D", "E501", "F841"]

[tool.pytest.ini_options]
testpaths = "endorser"
addopts = """
    --quiet --junitxml=./test-reports/junit.xml
    --cov-config .coveragerc --cov=endorser --cov-report term --cov-report xml
    --ruff
"""
markers = []
junit_family = "xunit1"
asyncio_mode = "auto"
filterwarnings = [
    'ignore:distutils Version classes are deprecated. Use packaging.version instead.:DeprecationWarning', # Ignore specific DeprecationWarning for old packages using distutils version class
]

[tool.coverage.run]

omit = [
    "*/tests/*",
    "demo/*",
    "docker/*",
    "docs/*",
    "scripts/*",
]
data_file = "test-reports/.coverage"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@abstract"
]
precision = 2
skip_covered = true
show_missing = true

[tool.coverage.xml]
output = "test-reports/coverage.xml"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
